+++ src/pages/match/[id]/present/index.tsx
@@ -4,91 +4,58 @@ import type { Session } from "next-auth";
 import { getSession } from "next-auth/react";
 import Head from "next/head";
 import type { ParsedUrlQuery } from "node:querystring";
-import { useState } from "react";
-import MatchScoreBox from "../../../../components/matchScoreBox";
-import { appRouter } from "../../../../server/api/root";
-import { api } from "../../../../utils/api";
-import { prisma } from "../../../../server/db";
-import { FullScreen, useFullScreenHandle } from "react-full-screen";
+import Header from "../../../../components/header";
+import presentScreens from "./list";
+import Image from "next/image";
+import Link from "next/link";
 
 const PresentScores: NextPage<{
-  match: Match & {
-    Team1: Team;
-    Team2: Team;
-  };
+  id: string;
   session: Session | null;
-}> = ({ match }) => {
-  const [scoreData, setScoreData] = useState<matchScore>({ events: [] });
-  api.scoring.fetch.useQuery(
-    { id: match.id },
-    {
-      onSuccess(data) {
-        if (data.data.status != "ok") throw new Error();
-        setScoreData(data.data.score);
-      },
-      refetchInterval: 300,
-    }
-  );
-
-  const handleFullScreen = useFullScreenHandle();
-  const [hideFullScreenBar, setHideFullScreenBar] = useState(false);
-  const handleHideFullScreenBar = () => {
-    setHideFullScreenBar(true);
-  };
+}> = ({ id }) => {
+  const items: JSX.Element[] = [];
+  for (const item of presentScreens) {
+    items.push(
+      <div
+        className="m-2 flex flex-col items-center justify-center p-1 dark:bg-[#f1fa8c]"
+        key={item.slug}
+      >
+        <Image
+          src={`/present-screenshots/${item.slug}.png`}
+          alt={`Screenshot of ${item.name}`}
+          title={item.description}
+          height={500}
+          width={500}
+        />
+        <div className="text-sm">
+          by{" "}
+          <a
+            href={item.authorLink}
+            className="text-blue-600 transition-colors dark:text-[#6272a4] dark:hover:text-[#44475a]"
+          >
+            {item.author}
+          </a>
+        </div>
+        <Link href={`/match/${id}/present/${item.slug}`}>
+          <button className="inline-block rounded border-2 border-blue-600 px-6 py-2 text-xs font-medium uppercase leading-tight text-blue-600 transition duration-150 ease-in-out hover:bg-black hover:bg-opacity-5 focus:outline-none focus:ring-0">
+            {item.name}
+          </button>
+        </Link>
+      </div>
+    );
+  }
 
   return (
     <>
       <Head>
         <title>Present Scores</title>
       </Head>
-      <FullScreen handle={handleFullScreen}>
-        <main className="flex h-screen max-h-screen w-screen flex-col justify-center">
-          <div className="grid grid-cols-2">
-            <div
-              className={`flex h-screen  flex-col items-center justify-center border border-slate-700`}
-              style={{
-                background: `linear-gradient(to top, ${
-                  match.Team1.color1 ?? "#FFFFFF"
-                } 0%, ${match.Team1.color2 ?? "#FFFFFF"} 100%)`,
-              }}
-            >
-              <div className="text-center text-[5vh]">{match.Team1.name}</div>
-              <MatchScoreBox data={scoreData} team={1} />
-            </div>
-            <div
-              className={`flex h-screen  flex-col items-center justify-center border border-slate-700`}
-              style={{
-                background: `linear-gradient(to top, ${
-                  match.Team2.color1 ?? "#FFFFFF"
-                } 0%, ${match.Team2.color2 ?? "#FFFFFF"} 100%)`,
-              }}
-            >
-              <div className="text-center text-[5vh]">{match.Team2.name}</div>
-              <MatchScoreBox data={scoreData} team={2} />
-            </div>
-          </div>
-        </main>
-        {!handleFullScreen.active && !hideFullScreenBar && (
-          <div className="flex flex-row justify-around dark:bg-[#282a36] dark:text-[#f8f8f2]">
-            <button
-              className="border border-blue-500 p-2 text-2xl dark:border-[#6272a4]"
-              onClick={() => {
-                void handleFullScreen.enter();
-              }}
-            >
-              enter fullscreen
-            </button>
-            <button
-              className="border border-blue-500 p-2 text-2xl dark:border-[#6272a4]"
-              onClick={() => {
-                handleHideFullScreenBar();
-              }}
-            >
-              hide this
-            </button>
-          </div>
-        )}
-      </FullScreen>
+      <div className="h-screen w-screen dark:bg-[#282a36]">
+        <Header />
+        <div className="flex flex-wrap justify-around dark:bg-[#282a36]">
+          {items}
+        </div>
+      </div>
     </>
   );
 };
@@ -102,29 +69,10 @@ export const getServerSideProps = async (
   const { id } = context.params!;
   if (!id || Array.isArray(id)) throw new Error("bad id");
 
-  const caller = appRouter.createCaller({
-    session: session,
-    prisma: prisma,
-  });
-
-  const match = await caller.match.fetch({ id });
-  if (match.data.status == "not-found")
-    return {
-      notFound: true,
-    };
-  if (match.data.status != "ok") throw new Error();
-
-  if (!match.data.match.Team1 || !match.data.match.Team2)
-    return {
-      redirect: {
-        destination: "/match/" + match.data.match.id,
-        permanent: false,
-      },
-    };
   return {
     props: {
       session,
-      match: match.data.match,
+      id,
     },
   };
 };
