+++ src/pages/match/[id]/present/list.ts
@@ -1,0 +1,11 @@
+const presentScreens: presentScreen[] = [
+  {
+    name: '"Classic"',
+    author: "Dominic",
+    authorLink: "https://github.com/mrhappyma",
+    description: "The original score screen",
+    slug: "classic",
+  },
+];
+
+export default presentScreens;
