+++ src/types/presentScreen.d.ts
@@ -1,0 +1,7 @@
+declare type presentScreen = {
+  name: string;
+  description: string;
+  author: string;
+  authorLink: string;
+  slug: string;
+};
